+++ packages/wonder-blocks-tokens/src/util/index.ts
@@ -1,4 +1,5 @@
 import {mix, fade} from "./color-utils";
 import {pxToRem, remToPx} from "./sizing-utils";
+import {tokenValue} from "./token-value";
 
-export {mix, fade, pxToRem, remToPx};
+export {mix, fade, pxToRem, remToPx, tokenValue};