+++ packages/wonder-blocks-pill/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @khanacademy/wonder-blocks-pill
 
+## 3.1.60
+
+### Patch Changes
+
+- Updated dependencies [40f217b]
+    - @khanacademy/wonder-blocks-tokens@16.3.0
+    - @khanacademy/wonder-blocks-clickable@8.1.7
+    - @khanacademy/wonder-blocks-link@10.1.9
+    - @khanacademy/wonder-blocks-styles@0.2.42
+    - @khanacademy/wonder-blocks-typography@4.3.1
+
+## 3.1.59
+
+### Patch Changes
+
+- 22cdea4: Fixed and unskipped tests
+
 ## 3.1.58
 
 ### Patch Changes